api file remove 경고창 확인

+++ public/js/book/form.js
@@ -23,9 +23,11 @@ if(document.querySelector('#btRemoveFile'))
 	document.querySelector('#btRemoveFile').addEventListener('click', onRemoveFile)
 
 function onRemoveFile(e) {
-	var idx = this.dataset['idx'];
-	var parent = this.parentNode;
-	axios.delete('/api/book/file/'+idx).then(onSuccess).catch(onError);
+	if(confirm('파일을 삭제하시겠습니까?\n삭제하신 파일은 되돌릴 수 없습니다.')) {
+		var idx = this.dataset['idx'];
+		var parent = this.parentNode;
+		axios.delete('/api/book/file/'+idx).then(onSuccess).catch(onError);
+	}
 	function onSuccess(r) {
 		console.log(r.data.code)
 		if(r.data.code == 200) parent.remove();
